+++ docs/python.md
@@ -21,3 +21,13 @@ To compile functions in the `func` dir we need to build the required libraries, 
 inv compile-libfaasm-emscripten
 inv funcs-emscripten
 ```
+
+## Running
+
+To run Python you need to set up the Faasm runtime root properly. To do this you can run:
+
+```
+sudo mkdir /usr/local/faasm/runtime_root
+sudo chown -R <user>:<user> !$
+inv set-up-python-root
+```
