+++ func/python/CMakeLists.txt
@@ -3,8 +3,8 @@ set(FAASM_USER python)
 # Normally we can rely on CMake to find the python libs, but when doing a
 # wasm build we need to point to the wasm-built versions on the local machine
 if (FAASM_BUILD_TYPE STREQUAL "emscripten")
-    set(PYTHON_LIBRARIES "/usr/local/code/cpython-emscripten/installs/python-3.5.2/lib/libpython3.5.a")
-    set(PYTHON_INCLUDE_DIRS "/usr/local/code/cpython-emscripten/installs/python-3.5.2/include/python3.5")
+    set(PYTHON_LIBRARIES "/home/shillaker/faasm/cpython-emscripten/installs/python-3.5.2/lib/libpython3.5.a")
+    set(PYTHON_INCLUDE_DIRS "/home/shillaker/faasm/cpython-emscripten/installs/python-3.5.2/include/python3.5")
 
     set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -s EXPORTED_FUNCTIONS='[\"_main\",\"_PyRun_SimpleString\"]'")
 else ()
