+++ tasks/__init__.py
@@ -1,7 +1,23 @@
+from tasks.aws import (
+    lambda_clear_queue,
+    lambda_concurrency,
+    lambda_worker_count,
+    list_lambdas,
+    list_event_sources,
+    invoke_lambda,
+    invoke_faasm_lambda,
+    deploy_faasm_lambda,
+    delete_lambda,
+    delete_faasm_lambda,
+    deploy_wasm_lambda_func,
+    deploy_native_lambda_func,
+    sqs_length,
+    purge_sqs,
+    invoke_lambda_worker,
+    prepare_lambda_workers,
+)
 from tasks.compile import (
-    build_python_host,
-    build_python,
-    build_python_emscripten,
+
     clean_build,
     funcs,
     funcs_emscripten,
@@ -11,7 +27,6 @@ from tasks.compile import (
     compile_eigen_emscripten,
     compile_libcurl,
 )
-
 from tasks.docker import (
     build_all,
     build_base,
@@ -34,35 +49,13 @@ from tasks.docker import (
     pull,
     push_base
 )
-
 from tasks.docker_compose import start_all, stop_all, restart_all
-
-from tasks.aws import (
-    lambda_clear_queue,
-    lambda_concurrency,
-    lambda_worker_count,
-    list_lambdas,
-    list_event_sources,
-    invoke_lambda,
-    invoke_faasm_lambda,
-    deploy_faasm_lambda,
-    delete_lambda,
-    delete_faasm_lambda,
-    deploy_wasm_lambda_func,
-    deploy_native_lambda_func,
-    sqs_length,
-    purge_sqs,
-    invoke_lambda_worker,
-    prepare_lambda_workers,
+from tasks.python import (
+    build_python_host,
+    build_python,
+    build_python_emscripten,
+    set_up_python_root,
 )
-
-from tasks.upload import (
-    upload_funcs,
-    upload_func,
-)
-
-from tasks.wasm2wast import wast
-
 from tasks.sgd import (
     begin_aws_svm,
     deploy_sgd_wasm_lambda,
@@ -74,3 +67,8 @@ from tasks.sgd import (
     reuters_prepare_aws,
     zero_sgd_native_lambda
 )
+from tasks.upload import (
+    upload_funcs,
+    upload_func,
+)
+from tasks.wasm2wast import wast
